# go: add faster RegBatch type